Rename PmpPlugin -> PmpPluginNapot, PmpPluginOld -> PmpPlugin
