+++ src/main/scala/vexriscv/demo/GenSecure.scala
@@ -41,7 +41,6 @@ object GenSecure extends App {
         ),
         new PmpPlugin(
           regions = 16,
-          granularity = 32,
           ioRange = _(31 downto 28) === 0xf
         ),
         new DecoderSimplePlugin(
